+++ ios/Sources/AutoLayoutView.swift
@@ -4,31 +4,35 @@ import UIKit
 
 /// Container for all RecyclerListView children. This will automatically remove all gaps and overlaps for GridLayouts with flexible spans.
 /// Note: This cannot work for masonry layouts i.e, pinterest like layout
-@objc class AutoLayoutView: UIView {
+@objc public class AutoLayoutView: UIView {
+    #if RCT_NEW_ARCH_ENABLED
+    @objc public var onBlankAreaEventHandler: ((CGFloat, CGFloat) -> Void)?
+    #endif
+    
     @objc(onBlankAreaEvent)
     var onBlankAreaEvent: RCTDirectEventBlock?
 
-    @objc func setHorizontal(_ horizontal: Bool) {
+    @objc public func setHorizontal(_ horizontal: Bool) {
         self.horizontal = horizontal
     }
 
-    @objc func setScrollOffset(_ scrollOffset: Int) {
+    @objc public func setScrollOffset(_ scrollOffset: Int) {
         self.scrollOffset = CGFloat(scrollOffset)
     }
 
-    @objc func setWindowSize(_ windowSize: Int) {
+    @objc public func setWindowSize(_ windowSize: Int) {
         self.windowSize = CGFloat(windowSize)
     }
 
-    @objc func setRenderAheadOffset(_ renderAheadOffset: Int) {
+    @objc public func setRenderAheadOffset(_ renderAheadOffset: Int) {
         self.renderAheadOffset = CGFloat(renderAheadOffset)
     }
 
-    @objc func setEnableInstrumentation(_ enableInstrumentation: Bool) {
+    @objc public func setEnableInstrumentation(_ enableInstrumentation: Bool) {
         self.enableInstrumentation = enableInstrumentation
     }
 
-    @objc func setDisableAutoLayout(_ disableAutoLayout: Bool) {
+    @objc public func setDisableAutoLayout(_ disableAutoLayout: Bool) {
         self.disableAutoLayout = disableAutoLayout
     }
 
@@ -39,17 +43,18 @@ import UIKit
     private var enableInstrumentation = false
     private var disableAutoLayout = false
 
+    /// Tracks where the last pixel is drawn in the overall
+    private var lastMaxBoundOverall: CGFloat = 0
     /// Tracks where the last pixel is drawn in the visible window
     private var lastMaxBound: CGFloat = 0
     /// Tracks where first pixel is drawn in the visible window
     private var lastMinBound: CGFloat = 0
 
-    override func layoutSubviews() {
+    override public func layoutSubviews() {
         fixLayout()
         super.layoutSubviews()
 
-        let scrollView = sequence(first: self, next: { $0.superview }).first(where: { $0 is UIScrollView })
-        guard enableInstrumentation, let scrollView = scrollView as? UIScrollView else { return }
+        guard enableInstrumentation, let scrollView = getScrollView() else { return }
 
         let scrollContainerSize = horizontal ? scrollView.frame.width : scrollView.frame.height
         let currentScrollOffset = horizontal ? scrollView.contentOffset.x : scrollView.contentOffset.y
@@ -67,13 +72,21 @@ import UIKit
             distanceFromWindowStart: distanceFromWindowStart,
             distanceFromWindowEnd: distanceFromWindowEnd
         )
-
+        
+        #if RCT_NEW_ARCH_ENABLED
+        onBlankAreaEventHandler?(blankOffsetStart, blankOffsetEnd)
+        #else
         onBlankAreaEvent?(
             [
                 "offsetStart": blankOffsetStart,
                 "offsetEnd": blankOffsetEnd,
             ]
         )
+        #endif
+    }
+
+    func getScrollView() -> UIScrollView? {
+        return sequence(first: self, next: { $0.superview }).first(where: { $0 is UIScrollView }) as? UIScrollView
     }
 
     /// Sorts views by index and then invokes clearGaps which does the correction.
@@ -86,8 +99,8 @@ import UIKit
             !disableAutoLayout
         else { return }
         let cellContainers = subviews
-            .compactMap { subview -> CellContainer? in
-                if let cellContainer = subview as? CellContainer {
+            .compactMap { subview -> CellContainerComponentView? in
+                if let cellContainer = subview as? CellContainerComponentView {
                     return cellContainer
                 } else {
                     assertionFailure("CellRendererComponent outer view should always be CellContainer. Learn more here: https://shopify.github.io/flash-list/docs/usage#cellrenderercomponent.")
@@ -96,16 +109,17 @@ import UIKit
             }
             .sorted(by: { $0.index < $1.index })
         clearGaps(for: cellContainers)
+        fixFooter()
     }
 
     /// Checks for overlaps or gaps between adjacent items and then applies a correction.
     /// Performance: RecyclerListView renders very small number of views and this is not going to trigger multiple layouts on the iOS side.
-    private func clearGaps(for cellContainers: [CellContainer]) {
+    private func clearGaps(for cellContainers: [CellContainerComponentView]) {
         var maxBound: CGFloat = 0
         var minBound: CGFloat = CGFloat(Int.max)
         var maxBoundNextCell: CGFloat = 0
         let correctedScrollOffset = scrollOffset - (horizontal ? frame.minX : frame.minY)
-
+        lastMaxBoundOverall = 0
         cellContainers.indices.dropLast().forEach { index in
             let cellContainer = cellContainers[index]
             let cellTop = cellContainer.frame.minY
@@ -115,67 +129,82 @@ import UIKit
 
             let nextCell = cellContainers[index + 1]
             let nextCellTop = nextCell.frame.minY
-            let nextCellBottom = nextCell.frame.maxY
             let nextCellLeft = nextCell.frame.minX
-            let nextCellRight = nextCell.frame.maxX
 
+			// Only apply correction if the next cell is consecutive.
+			let isNextCellConsecutive = nextCell.index == cellContainer.index + 1
+			
+			let isCellVisible = isWithinBounds(
+				cellContainer,
+				scrollOffset: correctedScrollOffset,
+				renderAheadOffset: renderAheadOffset,
+				windowSize: windowSize,
+				isHorizontal: horizontal
+			)
+			let isNextCellVisible = isWithinBounds(
+				nextCell,
+				scrollOffset: correctedScrollOffset,
+				renderAheadOffset: renderAheadOffset,
+				windowSize: windowSize,
+				isHorizontal: horizontal
+			)
 
             guard
-                isWithinBounds(
-                    cellContainer,
-                    scrollOffset: correctedScrollOffset,
-                    renderAheadOffset: renderAheadOffset,
-                    windowSize: windowSize,
-                    isHorizontal: horizontal
-                )
-            else { return }
-            let isNextCellVisible = isWithinBounds(
-                nextCell,
-                scrollOffset: correctedScrollOffset,
-                renderAheadOffset: renderAheadOffset,
-                windowSize: windowSize,
-                isHorizontal: horizontal
-            )
+                isCellVisible || isNextCellVisible
+            else {
+                updateLastMaxBoundOverall(currentCell: cellContainer, nextCell: nextCell)
+                return
+            }
+           
 
             if horizontal {
                 maxBound = max(maxBound, cellRight)
                 minBound = min(minBound, cellLeft)
                 maxBoundNextCell = maxBound
-                if cellTop < nextCellTop {
-                    if cellBottom != nextCellTop {
-                        nextCell.frame.origin.y = cellBottom
-                    }
-                    if cellLeft != nextCellLeft {
-                        nextCell.frame.origin.x = cellLeft
-                    }
-                } else {
-                    nextCell.frame.origin.x = maxBound
-                }
+				if isNextCellConsecutive {
+					if cellTop < nextCellTop {
+						if cellBottom != nextCellTop {
+							nextCell.frame.origin.y = cellBottom
+						}
+						if cellLeft != nextCellLeft {
+							nextCell.frame.origin.x = cellLeft
+						}
+					} else {
+						nextCell.frame.origin.x = maxBound
+					}
+				}
                 if isNextCellVisible {
-                    maxBoundNextCell = max(maxBound, nextCellRight)
+                    maxBoundNextCell = max(maxBound, nextCell.frame.maxX)
                 }
             } else {
                 maxBound = max(maxBound, cellBottom)
                 minBound = min(minBound, cellTop)
                 maxBoundNextCell = maxBound
-                if cellLeft < nextCellLeft {
-                    if cellRight != nextCellLeft {
-                        nextCell.frame.origin.x = cellRight
-                    }
-                    if cellTop != nextCellTop {
-                        nextCell.frame.origin.y = cellTop
-                    }
-                } else {
-                    nextCell.frame.origin.y = maxBound
-                }
+				if isNextCellConsecutive {
+					if cellLeft < nextCellLeft {
+						if cellRight != nextCellLeft {
+							nextCell.frame.origin.x = cellRight
+						}
+						if cellTop != nextCellTop {
+							nextCell.frame.origin.y = cellTop
+						}
+					} else {
+						nextCell.frame.origin.y = maxBound
+					}
+				}
                 if isNextCellVisible {
-                    maxBoundNextCell = max(maxBound, nextCellBottom)
+                    maxBoundNextCell = max(maxBound, nextCell.frame.maxY)
                 }
             }
+            updateLastMaxBoundOverall(currentCell: cellContainer, nextCell: nextCell)
         }
 
         lastMaxBound = maxBoundNextCell
         lastMinBound = minBound
+    }
+
+    private func updateLastMaxBoundOverall(currentCell: CellContainerComponentView, nextCell: CellContainerComponentView) {
+        lastMaxBoundOverall = max(lastMaxBoundOverall, horizontal ? currentCell.frame.maxX : currentCell.frame.maxY, horizontal ? nextCell.frame.maxX : nextCell.frame.maxY)
     }
 
     func computeBlankFromGivenOffset(
@@ -199,7 +228,7 @@ import UIKit
 
     /// It's important to avoid correcting views outside the render window. An item that isn't being recycled might still remain in the view tree. If views outside get considered then gaps between unused items will cause algorithm to fail.
     func isWithinBounds(
-        _ cellContainer: CellContainer,
+        _ cellContainer: CellContainerComponentView,
         scrollOffset: CGFloat,
         renderAheadOffset: CGFloat,
         windowSize: CGFloat,
@@ -214,5 +243,52 @@ import UIKit
         } else {
             return (cellFrame.minY >= boundsStart || cellFrame.maxY >= boundsStart) && (cellFrame.minY <= boundsEnd || cellFrame.maxY <= boundsEnd)
         }
+    }
+
+    /// Fixes footer position along with rest of the items
+    private func fixFooter() {
+        guard !disableAutoLayout, let parentScrollView = getScrollView() else {
+            return
+        }
+
+        let isAutoLayoutEndVisible = horizontal ? frame.maxX <= parentScrollView.frame.width : frame.maxY <= parentScrollView.frame.height
+        guard isAutoLayoutEndVisible, let footer = footer() else {
+            return
+        }
+
+        let diff = footerDiff()
+        guard diff != 0 else { return }
+
+        if horizontal {
+            footer.frame.origin.x += diff
+            frame.size.width += diff
+            superview?.frame.size.width += diff
+        } else {
+            footer.frame.origin.y += diff
+            frame.size.height += diff
+            superview?.frame.size.height += diff
+        }
+    }
+
+    private func footerDiff() -> CGFloat {
+        if subviews.count == 0 {
+            lastMaxBoundOverall = 0
+        } else if subviews.count == 1 {
+            let firstChild = subviews[0]
+            lastMaxBoundOverall = horizontal ? firstChild.frame.maxX : firstChild.frame.maxY
+        }
+        let autoLayoutEnd = horizontal ? frame.width : frame.height
+        return lastMaxBoundOverall - autoLayoutEnd
+    }
+
+    private func footer() -> UIView? {
+        // On the new arch, AutoLayoutView is wrapped with AutoLayoutViewComponentView, so we need to go up one more level
+        #if RCT_NEW_ARCH_ENABLED
+        let parentSubviews = superview?.superview?.subviews
+        #else
+        let parentSubviews = superview?.subviews
+        #endif
+
+        return parentSubviews?.first(where:{($0 as? CellContainerComponentView)?.index == -1})
     }
 }