+++ src/__tests__/helpers/mountFlashList.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { ListRenderItem, Text } from "react-native";
+import { Text } from "react-native";
 import "@quilted/react-testing/matchers";
-import { mount, Root } from "@quilted/react-testing";
+import { render, Root } from "@quilted/react-testing";
 
 import FlashList from "../../FlashList";
-import { FlashListProps } from "../../FlashListProps";
+import { FlashListProps, ListRenderItem } from "../../FlashListProps";
 
 jest.mock("../../FlashList", () => {
   const ActualFlashList = jest.requireActual("../../FlashList").default;
@@ -34,7 +34,7 @@ export const mountFlashList = (
   props?: MockFlashListProps,
   ref?: React.RefObject<FlashList<string>>
 ) => {
-  const flashList = mount(renderFlashList(props, ref)) as Omit<
+  const flashList = render(renderFlashList(props, ref)) as Omit<
     Root<FlashListProps<string>>,
     "instance"
   > & {